feat: implement logout
