+++ .mockery.yaml
@@ -17,3 +17,4 @@ packages:
       SensorRepository:
       TreeRepository:
       AuthRepository: 
+      UserRepository:
