+++ internal/server/http/entities/auth/logout.go
@@ -1,0 +1,5 @@
+package auth
+
+type LogoutRequest struct {
+	RefreshToken string `json:"refresh_token"`
+} // @Name LogoutRequest
