+++ internal/storage/auth/storage.go
@@ -8,10 +8,10 @@ import (
 
 func NewRepository(cfg *config.IdentityAuthConfig) *storage.Repository {
 	authRepo := keycloak.NewKeycloakRepository(cfg)
-  userRepo := keycloak.NewUserRepository(cfg)
+	userRepo := keycloak.NewUserRepository(cfg)
 
 	return &storage.Repository{
 		Auth: authRepo,
-    User: userRepo,
+		User: userRepo,
 	}
 }
